feat(authentication): add WebAuthn support and refactor login flow
Introduce WebAuthn as a new authentication method and refactor the existing login
flow to accommodate this change. The update includes modifications to the package
dependencies, context and service interfaces, and the implementation of new
utility functions to handle the authentication process.

BREAKING CHANGE: The login process now requires WebAuthn support in browsers and
the related API adjustments might affect existing clients integrating with the
use-ic-siwp-identity package.